Continued Work Downloading Articles
Articles/Feed Items now download when the user opens the app or links with Inoreader

+++ Library/WagsReaderLibrary/Inoreader/Models/UserToken.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using System;
 
 namespace WagsReaderLibrary.Inoreader.Models
 {
@@ -11,7 +12,7 @@ namespace WagsReaderLibrary.Inoreader.Models
         public string TokenType { get; set; }
 
         [JsonProperty("expires_in")]
-        public int ExpiresIn { get; set; }
+        public long ExpiresIn { get; set; }
 
         [JsonProperty("refresh_token")]
         public string RefreshToken { get; set; }
